+++ CHANGELOG.md
@@ -11,6 +11,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased][]
 
+### Added
+
+- Add dependency "[cake-case](https://github.com/archco/cake-case#readme)". #103
+
+### Deprecated
+
+- Some JS util methods are deprecated. use [`cake-case`](https://github.com/archco/cake-case#readme) instead.
+  - `Util.strToCamel`: use `Case.camel` or `Case.pascal` instead.
+  - `Util.strToKebab`: use `Case.kebab` instead.
+  - `Util.strToNormal`: use `Case.lower` instead.
+  - `Util.capitalize`: use `Case.capital` instead.
+
 ## [0.5.3][] - 2018-03-18
 
 ### Added